travel-assistant/pyproject.toml
[project]
name = "travel-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geonamescache>=2.0.0",
    "openai>=1.97.1",
    "openai-agents>=0.2.3",
    "pycountry>=24.6.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
